+++ src/pages/JobDetail/ui/MoreJob.jsx
@@ -1,11 +1,24 @@
 import React from 'react';
+import * as S from './style';
+import {styled} from "styled-components";
+import IntroArea from "@pages/JobDetail/components/IntroArea.jsx";
+import SiteArea from "@pages/JobDetail/components/SiteArea.jsx";
 
 const MoreJob = () => {
   return (
-      <div>
-rr
-      </div>
+      <Screen>
+        <IntroArea />
+        <SiteArea />
+      </Screen>
   );
 };
 
 export default MoreJob;
+
+const Screen = styled.div`
+  width: 100vw;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+`;
